Implement manual adding of select pages to llms.txt
